lazy load sign in gate component
